Frontend + improvements

+++ stock-price-analytics/src/main/java/com/leadsIq/stockPriceAnalytics/domain/entity/StockPrice.java
@@ -1,9 +1,7 @@
 package com.leadsIq.stockPriceAnalytics.domain.entity;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
-public record StockPrice(String company, LocalDate date, BigDecimal open, BigDecimal close,
-                         BigDecimal high, BigDecimal low, Long volume) {
+public record StockPrice(String company, LocalDate date, Double open, Double close, Double high, Double low, Long volume) {
 
 }
